requirements.txt
python-telegram-bot==22.2
langchain==0.2.0
langchain-chroma==0.1.2
sentence-transformers==2.3.1
chromadb==0.5.0
openai==1.10.0
openai-whisper==20231117
pydub==0.25.1
flask==3.0.0
pandas==2.1.4
aiofiles==0.8.0
python-dotenv==1.0.0
huggingface_hub==0.23.2
python-docx==1.1.0
pymupdf==1.24.4